Integrated delete

+++ RestApi/Sepes.RestApi/Controllers/StudyController.cs
@@ -28,64 +28,65 @@ namespace Sepes.RestApi.Controller
         [HttpPost("save")]
         public async Task<ActionResult<StudyInput>> SaveStudy([FromBody] StudyInput[] studies)
         {
-            //Begin sketch, this code needs to be integrated in bellow checks. Reason for keeping it outside studyservice would be to be able to return actionresults for error handling in frontend
-            //Error handling draft currently in [HttpPost("pod/delete")] and functional in there. 
             //Studies [1] is what the frontend claims the changes is based on while Studies [0] is the new version
-
-            //These must happen only if no changes are needed. Maybe must be moved to studyservice
-            List<ushort?> podIdBased = null;
-            List<ushort?> podIdNew = null;
-
-            foreach (PodInput podBased in studies[1].pods)
+            
+            //If based has more pods than new version then it can be assumed pods are requested to be deleted.
+            if (studies[1].pods.Count() > studies[0].pods.Count())
             {
-                podIdBased.Add(podBased.podId);
-            }
-            foreach (PodInput podNew in studies[0].pods)
-            {
-                podIdNew.Add(podNew.podId);
-            }
+                
+                List<ushort?> podIdBased = null;
+                List<ushort?> podIdNew = null;
 
-            var podIdDiff = podIdBased.Except(podIdNew);
-
-            //Make sure only one deletion has been ordered
-            if (podIdDiff.Count() == 1)
-            {
-                //Fetch the pod whose podId matches
-                foreach (PodInput podIn in studies[1].pods)
+                foreach (PodInput podBased in studies[1].pods)
                 {
-                    if (podIn.podId == podIdDiff.First())
-                    {
-                        //Perform delete
-                        Pod podToDelete = podIn.ToPod();
-                        var response = await _studyService.DeletePod(podToDelete);
+                    podIdBased.Add(podBased.podId);
+                }
+                foreach (PodInput podNew in studies[0].pods)
+                {
+                    podIdNew.Add(podNew.podId);
+                }
 
-                        if (response == podIn.podId)
+                var podIdDiff = podIdBased.Except(podIdNew);
+
+                //Make sure only one deletion has been ordered
+                if (podIdDiff.Count() == 1)
+                {
+                    //Fetch the pod whose podId matches
+                    foreach (PodInput podIn in studies[1].pods)
+                    {
+                        if (podIn.podId == podIdDiff.First())
                         {
-                            return Ok("Pod: " + podIn.podId + " deleted.");
-                        }
-                        else if (response == -1)
-                        {
-                            return BadRequest("Error: Pod does not match memory or is not found. Refresh and try again.");
-                        }
-                        else
-                        {
-                            return StatusCode(500, "Deletion status unknown. Refresh browser and check again.");
+                            //Perform delete
+                            Pod podToDelete = podIn.ToPod();
+                            var response = await _studyService.DeletePod(podToDelete);
+
+                            if (response == podIn.podId)
+                            {
+                                return Ok("Pod: " + podIn.podId + " deleted.");
+                            }
+                            else if (response == -1)
+                            {
+                                return BadRequest("Error: Pod does not match memory or is not found. Refresh and try again.");
+                            }
+                            else
+                            {
+                                return StatusCode(500, "Deletion status unknown. Refresh browser and check again.");
+                            }
                         }
                     }
                 }
+                else
+                {
+                    return BadRequest("Too many deletion requests. Only single deletions supported at this time.");
+                }
             }
-            else
-            {
-                return BadRequest("Too many deletion requests. Only single deletions supported at this time.");
-            }
-            //End sketch
-
-
-            if (studies[1] == null)
+            //If based is null it can be assumed this will be a newly created study
+            else if (studies[1] == null)
             {
                 Study study = await _studyService.Save(studies[0].ToStudy(), null);
                 return study.ToStudyInput();
             }
+            //Otherwise it must be a change.
             else
             {
                 Study study = await _studyService.Save(studies[0].ToStudy(), studies[1].ToStudy());
